fix: consolidation $6 param error, neural connection update, perception staleness
- semantic.js: cast episodeId to INT in contradiction insert (was null type error)
- neural-connections.js: remove unused $6 baseStrength param from UPDATE query
  (was left over after strength formula change, caused type inference failure)
- consolidation.js: per-item try/catch for connections (one bad connection
  no longer kills entire consolidation pass)
- cognitive-loop.js: fix neural bus pre-cycle using undefined currentLoad
  (replaced with getCognitiveLoad()), add error logging to catch blocks
- swift-bridge.js: build proper perception from cached events when shared
  state file is stale (fixes dashboard showing "unknown" for everything)

Made-with: Cursor

+++ memory/semantic.js
@@ -193,13 +193,13 @@ export async function contradict(conceptId, reason = '', {
   await pool.query(
     `INSERT INTO semantic_contradictions
      (concept_id, contradicting_concept_id, contradiction_set_id, episode_id, reason, weight)
-     VALUES ($1, $2, $3, $4, $5, $6)`,
+     VALUES ($1, $2, $3, $4::INT, $5, $6)`,
     [conceptId, contradictingConceptId, setId, episodeId, reason || null, clamp(Number(weight) || 1, 0.05, 5)]
   );
 
   await pool.query(
     `INSERT INTO semantic_evidence (concept_id, episode_id, source_type, weight, supports, evidence_text, metadata)
-     VALUES ($1, $2, 'contradiction', $3, false, $4, $5)`,
+     VALUES ($1, $2::INT, 'contradiction', $3, false, $4, $5)`,
     [conceptId, episodeId, clamp(Number(weight) || 1, 0.05, 5), reason || 'contradiction', JSON.stringify({ contradiction_set_id: setId })]
   );
